requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pymysql
passlib
python-jose[cryptography]
cachetools
